+++ Sources/Extensions/SwiftStdlib/ArrayExtensions.swift
@@ -280,21 +280,6 @@ public extension Array {
 		return slices
 	}
 
-	/// SwifterSwift: Group the elements of the array in a dictionary.
-	///
-	///     [0, 2, 5, 4, 7].groupByKey { $0%2 ? "evens" : "odds" } -> [ "evens" : [0, 2, 4], "odds" : [5, 7] ]
-	///
-	/// - Parameter getKey: Clousure to define the key for each element.
-	/// - Returns: A dictionary with values grouped with keys.
-	public func groupByKey<K: Hashable>(keyForValue: (_ element: Element) throws -> K) rethrows -> [K: [Element]] {
-		var group = [K: [Element]]()
-		for value in self {
-			let key = try keyForValue(value)
-			group[key] = (group[key] ?? []) + [value]
-		}
-		return group
-	}
-
 	/// SwifterSwift: Separates an array into 2 arrays based on a predicate.
 	///
 	///     [0, 1, 2, 3, 4, 5].divided { $0 % 2 == 0 } -> ( [0, 2, 4], [1, 3, 5] )